services/deepfake-detector/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
httpx==0.26.0
Pillow==10.2.0
numpy==1.26.3
google-generativeai==0.3.2
python-multipart==0.0.6
